+++ packages/expo-local-authentication/plugin/src/withLocalAuthentication.ts
@@ -3,7 +3,11 @@ import { AndroidConfig, ConfigPlugin, IOSConfig, createRunOncePlugin } from 'exp
 const pkg = require('expo-local-authentication/package.json');
 const FACE_ID_USAGE = 'Allow $(PRODUCT_NAME) to use Face ID';
 
-const withLocalAuthentication: ConfigPlugin<{ faceIDPermission?: string | false } | void> = (
+export type WithLocalAuthenticationProps = {
+  faceIDPermission?: string | false;
+};
+
+const withLocalAuthentication: ConfigPlugin<WithLocalAuthenticationProps | void> = (
   config,
   { faceIDPermission } = {}
 ) => {