+++ packages/expo-router/plugin/src/index.ts
@@ -3,6 +3,21 @@ import { validate } from 'schema-utils';
 
 const schema = require('../options.json');
 
+export type WithRouterProps = {
+  /** Production origin URL where assets in the public folder are hosted. The fetch function is polyfilled to support relative requests from this origin in production, development origin is inferred using the Expo CLI development server. */
+  origin?: string;
+  /** A more specific origin URL used in the `expo-router/head` module for iOS handoff. Defaults to `origin`. */
+  headOrigin?: string;
+  /** Should Async Routes be enabled. `production` is currently web-only and will be disabled on native. */
+  root?: string;
+  /** Should Async Routes be enabled, currently only `development` is supported. */
+  asyncRoutes?: string | { android?: string; ios?: string; web?: string; default?: string };
+  /** Should the sitemap be generated. Defaults to `true` */
+  sitemap?: boolean;
+  /** Generate partial typed routes */
+  partialTypedGroups?: boolean;
+};
+
 const withExpoHeadIos: ConfigPlugin = (config) => {
   return withInfoPlist(config, (config) => {
     // TODO: Add a way to enable this...
@@ -23,22 +38,7 @@ const withExpoHeadIos: ConfigPlugin = (config) => {
   });
 };
 
-const withRouter: ConfigPlugin<
-  {
-    /** Production origin URL where assets in the public folder are hosted. The fetch function is polyfilled to support relative requests from this origin in production, development origin is inferred using the Expo CLI development server. */
-    origin?: string;
-    /** A more specific origin URL used in the `expo-router/head` module for iOS handoff. Defaults to `origin`. */
-    headOrigin?: string;
-    /** Should Async Routes be enabled. `production` is currently web-only and will be disabled on native. */
-    root?: string;
-    /** Should Async Routes be enabled, currently only `development` is supported. */
-    asyncRoutes?: string | { android?: string; ios?: string; web?: string; default?: string };
-    /** Should the sitemap be generated. Defaults to `true` */
-    sitemap?: boolean;
-    /** Generate partial typed routes */
-    partialTypedGroups?: boolean;
-  } | void
-> = (config, _props) => {
+const withRouter: ConfigPlugin<WithRouterProps | void> = (config, _props) => {
   const props = _props || {};
   validate(schema, props);