+++ packages/expo-secure-store/plugin/src/withSecureStore.ts
@@ -12,12 +12,15 @@ const BACKUP_RULES_PATH = '@xml/secure_store_backup_rules';
 const EXTRACTION_RULES_PATH = '@xml/secure_store_data_extraction_rules';
 const FACEID_USAGE = 'Allow $(PRODUCT_NAME) to access your Face ID biometric data.';
 
-const withSecureStore: ConfigPlugin<
-  {
-    faceIDPermission?: string | false;
-    configureAndroidBackup?: boolean;
-  } | void
-> = (config, { faceIDPermission, configureAndroidBackup = true } = {}) => {
+export type WithSecureStoreProps = {
+  faceIDPermission?: string | false;
+  configureAndroidBackup?: boolean;
+};
+
+const withSecureStore: ConfigPlugin<WithSecureStoreProps | void> = (
+  config,
+  { faceIDPermission, configureAndroidBackup = true } = {}
+) => {
   IOSConfig.Permissions.createPermissionsPlugin({
     NSFaceIDUsageDescription: FACEID_USAGE,
   })(config, {