+++ packages/expo-tracking-transparency/plugin/src/withTrackingTransparency.ts
@@ -5,17 +5,20 @@ const pkg = require('expo-tracking-transparency/package.json');
 const DEFAULT_NSUserTrackingUsageDescription =
   'Allow this app to collect app-related data that can be used for tracking you or your device.';
 
-const withTrackingTransparency: ConfigPlugin<
-  {
-    /**
-     * Sets the iOS `NSUserTrackingUsageDescription` permission message in `Info.plist`. Omitting a
-     * description will result in using the default permission message.
-     * @default 'Allow this app to collect app-related data that can be used for tracking you or your
-     * device.'
-     */
-    userTrackingPermission?: string | false;
-  } | void
-> = (config, props) => {
+export type WithTrackingTransparencyProps = {
+  /**
+   * Sets the iOS `NSUserTrackingUsageDescription` permission message in `Info.plist`. Omitting a
+   * description will result in using the default permission message.
+   * @default 'Allow this app to collect app-related data that can be used for tracking you or your
+   * device.'
+   */
+  userTrackingPermission?: string | false;
+};
+
+const withTrackingTransparency: ConfigPlugin<WithTrackingTransparencyProps | void> = (
+  config,
+  props
+) => {
   IOSConfig.Permissions.createPermissionsPlugin({
     NSUserTrackingUsageDescription: DEFAULT_NSUserTrackingUsageDescription,
   })(config, {