+++ packages/expo-video/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 💡 Others
 
+- export config plugin types
+
 ## 2.0.0 — 2024-11-11
 
 ### 💡 Others